Install .pyi files
Add rules to generate (as part of the build) and install type stubs for
pydrake. Modify wheel build to include these as well.

Note that this makes numpy and matplotlib build-time dependencies.
(Previously they were only required to test/run pydrake.)

Note also that there are issues remaining with the .pyi's; in
particular, they are completely unusable by mypy.
