+++ bindings/pydrake/BUILD.bazel
@@ -24,6 +24,7 @@ load(
     "get_pybind_package_info",
 )
 load("//bindings/pydrake:pydrake.bzl", "add_lint_tests_pydrake")
+load("//bindings/pydrake:stubgen.bzl", "generate_python_stubs")
 
 package(default_visibility = [
     "//bindings:__subpackages__",
@@ -357,18 +358,6 @@ drake_py_library(
     ],
 )
 
-install(
-    name = "install",
-    install_tests = [
-        ":test/all_install_test.py",
-        ":test/common_install_test.py",
-    ],
-    targets = PY_LIBRARIES + [":all_py"],
-    py_dest = PACKAGE_INFO.py_dest,
-    visibility = ["//visibility:public"],
-    deps = get_drake_py_installs(PY_LIBRARIES_WITH_INSTALL),
-)
-
 drake_cc_googletest(
     name = "documentation_pybind_test",
     deps = [
@@ -685,22 +674,153 @@ generate_pybind_coverage(
 drake_py_binary(
     name = "stubgen",
     srcs = ["stubgen.py"],
-    args = ["--package=pydrake"],
     deps = [
         ":all_py",
         "@mypy_internal//:mypy",
     ],
 )
 
-# TODO(mwoehlke-kitware): actually install the .pyi's.
+# TODO(jwnimmer-tri): This list should be automatically generated.
+PYI_FILES = [
+    "pydrake/__init__.pyi",
+    "pydrake/all.pyi",
+    "pydrake/autodiffutils.pyi",
+    "pydrake/common/__init__.pyi",
+    "pydrake/common/_module_py.pyi",
+    "pydrake/common/all.pyi",
+    "pydrake/common/compatibility.pyi",
+    "pydrake/common/containers.pyi",
+    "pydrake/common/cpp_param.pyi",
+    "pydrake/common/cpp_template.pyi",
+    "pydrake/common/deprecation.pyi",
+    "pydrake/common/eigen_geometry.pyi",
+    "pydrake/common/jupyter.pyi",
+    "pydrake/common/pybind11_version.pyi",
+    "pydrake/common/schema.pyi",
+    "pydrake/common/value.pyi",
+    "pydrake/common/yaml.pyi",
+    "pydrake/examples/__init__.pyi",
+    "pydrake/examples/_examples_extra.pyi",
+    "pydrake/examples/acrobot.pyi",
+    "pydrake/examples/compass_gait.pyi",
+    "pydrake/examples/manipulation_station.pyi",
+    "pydrake/examples/multibody/__init__.pyi",
+    "pydrake/examples/multibody/cart_pole_passive_simulation.pyi",
+    "pydrake/examples/multibody/pendulum_lqr_monte_carlo_analysis.pyi",
+    "pydrake/examples/multibody/run_planar_scenegraph_visualizer.pyi",
+    "pydrake/examples/pendulum.pyi",
+    "pydrake/examples/quadrotor.pyi",
+    "pydrake/examples/rimless_wheel.pyi",
+    "pydrake/examples/van_der_pol.pyi",
+    "pydrake/forwarddiff.pyi",
+    "pydrake/geometry.pyi",
+    "pydrake/lcm.pyi",
+    "pydrake/manipulation/__init__.pyi",
+    "pydrake/manipulation/all.pyi",
+    "pydrake/manipulation/kuka_iiwa.pyi",
+    "pydrake/manipulation/planner.pyi",
+    "pydrake/manipulation/schunk_wsg.pyi",
+    "pydrake/math.pyi",
+    "pydrake/multibody/__init__.pyi",
+    "pydrake/multibody/all.pyi",
+    "pydrake/multibody/benchmarks.pyi",
+    "pydrake/multibody/inverse_kinematics.pyi",
+    "pydrake/multibody/jupyter_widgets.pyi",
+    "pydrake/multibody/math.pyi",
+    "pydrake/multibody/meshcat.pyi",
+    "pydrake/multibody/optimization.pyi",
+    "pydrake/multibody/parsing.pyi",
+    "pydrake/multibody/plant.pyi",
+    "pydrake/multibody/tree.pyi",
+    "pydrake/perception.pyi",
+    "pydrake/polynomial.pyi",
+    "pydrake/solvers/__init__.pyi",
+    "pydrake/solvers/_extra.pyi",
+    "pydrake/solvers/all.pyi",
+    "pydrake/solvers/augmented_lagrangian.pyi",
+    "pydrake/solvers/branch_and_bound.pyi",
+    "pydrake/solvers/clp.pyi",
+    "pydrake/solvers/csdp.pyi",
+    "pydrake/solvers/dreal.pyi",
+    "pydrake/solvers/gurobi.pyi",
+    "pydrake/solvers/ibex.pyi",
+    "pydrake/solvers/ipopt.pyi",
+    "pydrake/solvers/mathematicalprogram.pyi",
+    "pydrake/solvers/mixed_integer_optimization_util.pyi",
+    "pydrake/solvers/mixed_integer_rotation_constraint.pyi",
+    "pydrake/solvers/mosek.pyi",
+    "pydrake/solvers/nlopt.pyi",
+    "pydrake/solvers/osqp.pyi",
+    "pydrake/solvers/scs.pyi",
+    "pydrake/solvers/sdpa_free_format.pyi",
+    "pydrake/solvers/snopt.pyi",
+    # "pydrake/stubgen.pyi", Don't include this; why is it even generated?
+    "pydrake/symbolic.pyi",
+    "pydrake/systems/__init__.pyi",
+    "pydrake/systems/_resample_interp1d.pyi",
+    "pydrake/systems/all.pyi",
+    "pydrake/systems/analysis.pyi",
+    "pydrake/systems/controllers.pyi",
+    "pydrake/systems/drawing.pyi",
+    "pydrake/systems/framework.pyi",
+    "pydrake/systems/jupyter_widgets.pyi",
+    "pydrake/systems/lcm.pyi",
+    "pydrake/systems/meshcat_visualizer.pyi",
+    "pydrake/systems/perception.pyi",
+    "pydrake/systems/planar_scenegraph_visualizer.pyi",
+    "pydrake/systems/primitives.pyi",
+    "pydrake/systems/pyplot_visualizer.pyi",
+    "pydrake/systems/rendering.pyi",
+    "pydrake/systems/scalar_conversion.pyi",
+    "pydrake/systems/sensors.pyi",
+    "pydrake/systems/trajectory_optimization.pyi",
+    "pydrake/trajectories.pyi",
+    "pydrake/tutorials.pyi",
+    "pydrake/visualization/__init__.pyi",
+    "pydrake/visualization/all.pyi",
+    "pydrake/visualization/meldis.pyi",
+    "pydrake/visualization/plotting.pyi",
+]
+
+# TODO(mwoehlke-kitware): genrule inappropriately gives execute permission to
+# all its outputs; see https://github.com/bazelbuild/bazel/issues/3359.
+# (Applies to both :pydrake_pyi and, below, :pydrake_typed.)
+generate_python_stubs(
+    name = "pydrake_pyi",
+    outs = PYI_FILES,
+    package = "pydrake",
+    tool = ":stubgen",
+)
+
+# PEP 561 marker file; tells tools that type information is available.
+genrule(
+    name = "pydrake_typed",
+    srcs = [],
+    outs = ["pydrake/py.typed"],
+    cmd = "echo '# Marker file for PEP 561.' > $@",
+)
+
+install(
+    name = "install",
+    install_tests = [
+        ":test/all_install_test.py",
+        ":test/common_install_test.py",
+    ],
+    targets = PY_LIBRARIES + [":all_py"],
+    py_dest = PACKAGE_INFO.py_dest,
+    data = [
+        ":pydrake_pyi",
+        ":pydrake_typed",
+    ],
+    data_dest = "@PYTHON_SITE_PACKAGES@",
+    visibility = ["//visibility:public"],
+    deps = get_drake_py_installs(PY_LIBRARIES_WITH_INSTALL),
+)
 
 drake_py_unittest(
-    name = "stubgen_test",
-    timeout = "moderate",
-    deps = [
-        ":all_py",
-        "@mypy_internal//:mypy",
-    ],
+    name = "pyi_sanity_test",
+    data = [":pydrake_pyi"],
+    deps = ["@bazel_tools//tools/python/runfiles"],
 )
 
 add_lint_tests_pydrake(
