+++ tools/wheel/test/test-wheel.sh
@@ -56,3 +56,16 @@ pypi_wheel_max_size = 100 << 20  # 100 MiB
 fail_message = f"Wheel is too large ({wheel_size_in_mib:.2f} MiB) for PyPI"
 assert wheel_size < pypi_wheel_max_size, fail_message
 EOF
+
+python - "$1" << EOF
+import os
+import pydrake
+
+def assert_exists(path):
+    assert os.path.exists(path), f"{path!r} does not exist!"
+
+# Check that type information files are present.
+pydrake_dir = pydrake.__path__[0]
+assert_exists(os.path.join(pydrake_dir, 'py.typed'))
+assert_exists(os.path.join(pydrake_dir, 'all.pyi'))
+EOF
